Created types/protocol.py
Moved protocol collection to new file because of new calc prop,
removed it from __init__ collections. Moved experiment_type
collection to __init__ collection due to lack of calc props
