+++ src/encoded/types/__init__.py
@@ -109,6 +109,36 @@ class Enzyme(Item):
 
 
 @collection(
+    name='experiment-types',
+    unique_key='experiment_type:experiment_name',
+    lookup_key='title',
+    properties={
+        'title': 'Experiment Types',
+        'description': 'Listing of experiment types for 4DN items',
+    }
+)
+class ExperimentType(Item):
+    """The ExperimentType class that descrbes an experiment type that can be associated with an experiment."""
+
+    item_type = 'experiment_type'
+    schema = load_schema('encoded:schemas/experiment_type.json')
+    name_key = 'experiment_name'
+
+    embedded_list = Item.embedded_list + [
+        "sop.description",
+        "reference_pubs.short_attribution",
+        "reference_pubs.authors",
+        "reference_pubs.date_published",
+        "reference_pubs.journal"
+    ]
+
+    def _update(self, properties, sheets=None):
+        # set name based on what is entered into title
+        properties['experiment_name'] = set_namekey_from_title(properties)
+        super(ExperimentType, self)._update(properties, sheets)
+
+
+@collection(
     name='file-formats',
     unique_key='file_format:file_format',
     lookup_key='file_format',
@@ -218,54 +248,54 @@ class Organism(Item):
         return name
 
 
-@collection(
-    name='protocols',
-    properties={
-        'title': 'Protocols',
-        'description': 'Listing of protocols',
-    })
-class Protocol(Item, ItemWithAttachment):
-    """Protocol class."""
-
-    item_type = 'protocol'
-    schema = load_schema('encoded:schemas/protocol.json')
-    embedded_list = Item.embedded_list + ["award.project", "lab.title"]
-    rev = {
-        'exp_type': ('ExperimentType', 'other_protocols'),
-        'sop_exp': ('ExperimentType', 'sop')
-    }
-
-    @calculated_property(schema={
-        "title": "Display Title",
-        "description": "A calculated title",
-        "type": "string"
-    })
-    def display_title(self, protocol_type, title=None, attachment=None, date_created=None):
-        if title:
-            return title
-        elif attachment:
-            return attachment.get('download')
-        else:
-            if protocol_type == 'Other':
-                protocol_type = 'Protocol'
-            if date_created:  # pragma: no cover should always have this value
-                protocol_type = protocol_type + " from " + date_created[:10]
-            return protocol_type
-
-    @calculated_property(schema={
-        "title": "Experiment Type",
-        "description": "The type of experiment associated with this protocol",
-        "type": "string",
-        "linkTo": "ExperimentType"
-    })
-    def experiment_type(self, request):
-        sop_paths = self.rev_link_atids(request, 'sop_exp')
-        if sop_paths:
-            return sop_paths[0]
-        exptype_paths = self.rev_link_atids(request, 'exp_type')
-        if exptype_paths:
-            return exptype_paths[0]
-
+# @collection(
+#     name='protocols',
+#     properties={
+#         'title': 'Protocols',
+#         'description': 'Listing of protocols',
+#     })
+# class Protocol(Item, ItemWithAttachment):
+#     """Protocol class."""
+#
+#     item_type = 'protocol'
+#     schema = load_schema('encoded:schemas/protocol.json')
+#     embedded_list = Item.embedded_list + ["award.project", "lab.title"]
+#     rev = {
+#         'exp_type': ('ExperimentType', 'other_protocols'),
+#         'sop_exp': ('ExperimentType', 'sop')
+#     }
+#
+#     @calculated_property(schema={
+#         "title": "Display Title",
+#         "description": "A calculated title",
+#         "type": "string"
+#     })
+#     def display_title(self, protocol_type, title=None, attachment=None, date_created=None):
+#         if title:
+#             return title
+#         elif attachment:
+#             return attachment.get('download')
+#         else:
+#             if protocol_type == 'Other':
+#                 protocol_type = 'Protocol'
+#             if date_created:  # pragma: no cover should always have this value
+#                 protocol_type = protocol_type + " from " + date_created[:10]
+#             return protocol_type
+#
+#     @calculated_property(schema={
+#         "title": "Experiment Type",
+#         "description": "The type of experiment associated with this protocol",
+#         "type": "string",
+#         "linkTo": "ExperimentType"
+#     })
+#     def experiment_type(self, request):
+#         sop_paths = self.rev_link_atids(request, 'sop_exp')
+#         if sop_paths:
+#             return sop_paths[0]
+#         exptype_paths = self.rev_link_atids(request, 'exp_type')
+#         if exptype_paths:
+#             return exptype_paths[0]
+#
 
 @collection(
     name='tracking-items',
